+++ src/main/java/com/epam/reportportal/cucumber/RunningContext.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 EPAM Systems
+ * Copyright 2020 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -40,29 +40,33 @@ import java.util.stream.IntStream;
  *
  * @author Serhii Zharskyi
  * @author Vitaliy Tsvihun
+ * @author Vadzim Hushchanskou
  */
-class RunningContext {
+public class RunningContext {
 
 	private RunningContext() {
 		throw new AssertionError("No instances should exist for the class!");
 	}
 
-	static class FeatureContext {
+	public static class FeatureContext {
 		private static final Map<String, TestSourceRead> PATH_TO_READ_EVENT_MAP = new ConcurrentHashMap<>();
-		private String currentFeatureUri;
+		private final String currentFeatureUri;
+		private final Feature currentFeature;
+		private final Set<ItemAttributesRQ> attributes;
 		private Maybe<String> currentFeatureId;
-		private Feature currentFeature;
-		private Set<ItemAttributesRQ> attributes;
 
-		FeatureContext() {
-			attributes = new HashSet<>();
+		public FeatureContext(TestCase testCase) {
+			TestSourceRead event = PATH_TO_READ_EVENT_MAP.get(testCase.getUri());
+			currentFeature = getFeature(event.source);
+			currentFeatureUri = event.uri;
+			attributes = Utils.extractAttributes(currentFeature.getTags());
 		}
 
-		static void addTestSourceReadEvent(String path, TestSourceRead event) {
+		public static void addTestSourceReadEvent(String path, TestSourceRead event) {
 			PATH_TO_READ_EVENT_MAP.put(path, event);
 		}
 
-		ScenarioContext getScenarioContext(TestCase testCase) {
+		public ScenarioContext getScenarioContext(TestCase testCase) {
 			ScenarioDefinition scenario = getScenario(testCase);
 			ScenarioContext context = new ScenarioContext();
 			context.processTags(testCase.getTags());
@@ -70,18 +74,11 @@ class RunningContext {
 			context.setTestCase(testCase);
 			context.processBackground(getBackground());
 			context.processScenarioOutline(scenario);
+			context.setFeatureUri(getUri());
 			return context;
 		}
 
-		FeatureContext processTestSourceReadEvent(TestCase testCase) {
-			TestSourceRead event = PATH_TO_READ_EVENT_MAP.get(testCase.getUri());
-			currentFeature = getFeature(event.source);
-			currentFeatureUri = event.uri;
-			attributes = Utils.extractAttributes(currentFeature.getTags());
-			return this;
-		}
-
-		Feature getFeature(String source) {
+		public Feature getFeature(String source) {
 			Parser<GherkinDocument> parser = new Parser<>(new AstBuilder());
 			TokenMatcher matcher = new TokenMatcher();
 			GherkinDocument gherkinDocument;
@@ -94,33 +91,33 @@ class RunningContext {
 			return gherkinDocument.getFeature();
 		}
 
-		Background getBackground() {
+		public Background getBackground() {
 			ScenarioDefinition background = getFeature().getChildren().get(0);
 			return background instanceof Background ? (Background) background : null;
 		}
 
-		Feature getFeature() {
+		public Feature getFeature() {
 			return currentFeature;
 		}
 
-		Set<ItemAttributesRQ> getAttributes() {
+		public Set<ItemAttributesRQ> getAttributes() {
 			return attributes;
 		}
 
-		String getUri() {
+		public String getUri() {
 			return currentFeatureUri;
 		}
 
-		Maybe<String> getFeatureId() {
+		public Maybe<String> getFeatureId() {
 			return currentFeatureId;
 		}
 
-		void setFeatureId(Maybe<String> featureId) {
+		public void setFeatureId(Maybe<String> featureId) {
 			this.currentFeatureId = featureId;
 		}
 
 		@SuppressWarnings("unchecked")
-		<T extends ScenarioDefinition> T getScenario(TestCase testCase) {
+		public <T extends ScenarioDefinition> T getScenario(TestCase testCase) {
 			List<ScenarioDefinition> featureScenarios = getFeature().getChildren();
 			for (ScenarioDefinition scenario : featureScenarios) {
 				if (scenario instanceof Background) {
@@ -144,8 +141,9 @@ class RunningContext {
 		}
 	}
 
-	static class ScenarioContext {
+	public static class ScenarioContext {
 		private static final Map<ScenarioDefinition, List<Integer>> scenarioOutlineMap = new ConcurrentHashMap<>();
+
 		private final Queue<Step> backgroundSteps = new ArrayDeque<>();
 		private final Map<Integer, Step> scenarioLocationMap = new HashMap<>();
 		private Set<ItemAttributesRQ> attributes = new HashSet<>();
@@ -157,20 +155,18 @@ class RunningContext {
 		private ScenarioDefinition scenario;
 		private TestCase testCase;
 		private boolean hasBackground = false;
-		private String scenarioDesignation;
 		private String outlineIteration;
+		private String uri;
+		private String text;
 
-		ScenarioContext() {
-		}
-
-		void processScenario(ScenarioDefinition scenario) {
+		public void processScenario(ScenarioDefinition scenario) {
 			this.scenario = scenario;
 			for (Step step : scenario.getSteps()) {
 				scenarioLocationMap.put(step.getLocation().getLine(), step);
 			}
 		}
 
-		void processBackground(Background background) {
+		public void processBackground(Background background) {
 			if (background != null) {
 				this.background = background;
 				hasBackground = true;
@@ -186,7 +182,7 @@ class RunningContext {
 		/**
 		 * Takes the serial number of scenario outline and links it to the executing scenario
 		 **/
-		void processScenarioOutline(ScenarioDefinition scenarioOutline) {
+		public void processScenarioOutline(ScenarioDefinition scenarioOutline) {
 			if (isScenarioOutline(scenarioOutline)) {
 				scenarioOutlineMap.computeIfAbsent(scenarioOutline,
 						k -> ((ScenarioOutline) scenarioOutline).getExamples()
@@ -199,83 +195,82 @@ class RunningContext {
 						.filter(i -> getLine() == scenarioOutlineMap.get(scenarioOutline).get(i))
 						.findFirst()
 						.orElseThrow(() -> new IllegalStateException(String.format("No outline iteration number found for scenario %s",
-								scenarioDesignation
+								Utils.getCodeRef(uri, getLine())
 						)));
 				outlineIteration = String.format("[%d]", iterationIdx + 1);
 			}
 		}
 
-		void processTags(List<PickleTag> pickleTags) {
+		public void processTags(List<PickleTag> pickleTags) {
 			attributes = Utils.extractPickleTags(pickleTags);
 		}
 
-		void mapBackgroundSteps(Background background) {
+		public void mapBackgroundSteps(Background background) {
 			for (Step step : background.getSteps()) {
 				scenarioLocationMap.put(step.getLocation().getLine(), step);
 			}
 		}
 
-		String getName() {
+		public String getName() {
 			return scenario.getName();
 		}
 
-		String getKeyword() {
+		public String getKeyword() {
 			return scenario.getKeyword();
 		}
 
-		int getLine() {
+		public int getLine() {
 			return isScenarioOutline(scenario) ? testCase.getLine() : scenario.getLocation().getLine();
 		}
 
-		String getStepPrefix() {
+		public String getStepPrefix() {
 			return hasBackground() && withBackground() ? background.getKeyword().toUpperCase() + AbstractReporter.COLON_INFIX : "";
 		}
 
-		Step getStep(TestStep testStep) {
+		public Step getStep(TestStep testStep) {
 			PickleStepTestStep pickleStepTestStep = (PickleStepTestStep) testStep;
 			Step step = scenarioLocationMap.get(pickleStepTestStep.getStepLine());
 			if (step != null) {
 				return step;
 			}
-			throw new IllegalStateException(String.format("Trying to get step for unknown line in feature. " + "Scenario: %s, line: %s",
+			throw new IllegalStateException(String.format("Trying to get step for unknown line in feature. Scenario: %s, line: %s",
 					scenario.getName(),
 					getLine()
 			));
 		}
 
-		Maybe<String> getId() {
+		public Maybe<String> getId() {
 			return id;
 		}
 
-		void setId(Maybe<String> newId) {
+		public void setId(Maybe<String> newId) {
 			if (id != null) {
 				throw new IllegalStateException("Attempting re-set scenario ID for unfinished scenario: " + getName());
 			}
 			id = newId;
 		}
 
-		void setTestCase(TestCase testCase) {
+		public void setTestCase(TestCase testCase) {
 			this.testCase = testCase;
-			scenarioDesignation = testCase.getScenarioDesignation();
 		}
 
-		void nextBackgroundStep() {
+		public void nextBackgroundStep() {
 			backgroundSteps.poll();
 		}
 
-		boolean isScenarioOutline(ScenarioDefinition scenario) {
+		public boolean isScenarioOutline(ScenarioDefinition scenario) {
 			return scenario instanceof ScenarioOutline;
 		}
 
-		boolean withBackground() {
+		public boolean withBackground() {
 			return !backgroundSteps.isEmpty();
 		}
 
-		boolean hasBackground() {
+		public boolean hasBackground() {
 			return hasBackground && background != null;
 		}
 
-		String getOutlineIteration() {
+		public String getOutlineIteration() {
 			return outlineIteration;
 		}
 
@@ -301,6 +296,22 @@ class RunningContext {
 
 		public void setHookStatus(Result.Type hookStatus) {
 			this.hookStatus = hookStatus;
+		}
+
+		public void setFeatureUri(String featureUri) {
+			this.uri = featureUri;
+		}
+
+		public String getFeatureUri() {
+			return uri;
+		}
+
+		public void setCurrentText(String stepText) {
+			this.text = stepText;
+		}
+
+		public String getCurrentText() {
+			return text;
 		}
 	}
 }